Synced core libraries and examples with CI40 LMC

+++ include/click/accel.h
@@ -2,9 +2,9 @@
 #define H_CLICK_ACCEL
 
 
-uint8_t accel_click_enable();
-uint8_t accel_click_get_measure(float * accelX, float * accelY, float * accelZ);
-uint8_t accel_click_disable();
+int accel_click_enable();
+int accel_click_get_measure(float * accelX, float * accelY, float * accelZ);
+int accel_click_disable();
 
 
 #endif
